+++ lib/data/local/app_style.dart
@@ -1,0 +1,1 @@
+enum AppStyle { material, materialDynamic }