+++ lighthouse_provider/lighthouse_back_end/lib/src/ble/bluetooth_characteristic.dart
@@ -53,5 +53,4 @@ abstract class LHBluetoothCharacteristic {
     return await write(list, withoutResponse: withoutResponse);
   }
 //end region
-
 }